% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{dbReadModel}
\alias{dbReadModel}
\title{Load the disease spread model from the database}
\usage{
dbReadModel(dbname = "./model.sqlite")
}
\arguments{
\item{dbname}{The path to the database file. Default is
'./model.sqlite'.}
}
\description{
Load the disease spread model from the database
}
