% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{init}
\alias{init}
\title{Initialize an FMD model}
\usage{
init(dbname = "./model.sqlite", beta = 0.05, gamma = 0.077)
}
\arguments{
\item{dbname}{The path to the database file. Default is
'./model.sqlite'.}

\item{beta}{the transmission rate parameter.}

\item{gamma}{the recovery rate parameter.}
}
\description{
Initialize an FMD model
}
