% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run}
\alias{run}
\title{Simulate one time-step of the disease spread model}
\usage{
run(dbname = "./model.sqlite")
}
\arguments{
\item{dbname}{The path to the database file. Default is
'./model.sqlite'.}
}
\description{
Simulate one time-step of the disease spread model
}
